// kservices_ts_api/src/api/domain/entities/v3/ivr.entity.ts
export type updateAgentRequest = {
  sme_id: number;
  in_flag: number;
  in_call_id: string;
  in_agent_id: number;
  in_customer_ani: string;
  in_duration: number;
  in_filename: string;
  in_status: number;
  in_flag_tbl: string;
  in_ip: number;
  insertDateTime: string;
};

export type failedRecordingRequest = {
  sme_id: number;
  in_flag: number;
  in_call_id: string;
  in_agent_id: number;
  in_customer_ani: string;
  in_duration: number;
  in_filename: string;
  in_status: number;
  in_flag_tbl: string;
  in_ip: number;
  in_session_id: string;
  insertDateTime: string;
};

export type getFailedRecordingRequest = { 
  in_ip: string;
  total_records: number;
};

export type updateFailedRecordingRequest = { 
  in_call_id: string;
  in_sme_id: number;
  in_status: string;
  in_s3url: string;
  
};